chore: rename testhelpers.spec to testHelpers
